+++ odoo/local-src/connector_esb/tests/__init__.py
@@ -1,4 +1,5 @@
 from . import test_export_pharmacy
 from . import test_export_product
 from . import test_ws_product_stock
+from . import test_export_stock
 from . import test_xml_writer